requirements.txt
openai>=1.45.0
rapidocr-onnxruntime>=1.3.3
onnxruntime>=1.17.0
pandas>=2.1.0
python-dotenv>=1.0.1
rapidfuzz>=3.9.4
numpy>=1.24.0
